feat: 蓝盾对接权限中心RBAC #7794
