+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/common/MockAuthCoreAutoConfiguration.kt
@@ -14,6 +14,7 @@ import com.tencent.devops.auth.service.iam.PermissionItsmCallbackService
 import com.tencent.devops.auth.service.iam.PermissionProjectService
 import com.tencent.devops.auth.service.iam.PermissionResourceGroupService
 import com.tencent.devops.auth.service.iam.PermissionResourceService
+import com.tencent.devops.auth.service.iam.PermissionResourceValidateService
 import com.tencent.devops.auth.service.iam.PermissionRoleMemberService
 import com.tencent.devops.auth.service.iam.PermissionRoleService
 import com.tencent.devops.auth.service.iam.PermissionService
@@ -29,6 +30,7 @@ import com.tencent.devops.auth.service.sample.SamplePermissionGradeService
 import com.tencent.devops.auth.service.sample.SamplePermissionItsmCallbackService
 import com.tencent.devops.auth.service.sample.SamplePermissionResourceGroupService
 import com.tencent.devops.auth.service.sample.SamplePermissionResourceService
+import com.tencent.devops.auth.service.sample.SamplePermissionResourceValidateService
 import com.tencent.devops.auth.service.sample.SamplePermissionRoleMemberService
 import com.tencent.devops.auth.service.sample.SamplePermissionRoleService
 import com.tencent.devops.auth.service.sample.SamplePermissionUrlServiceImpl
@@ -105,4 +107,8 @@ class MockAuthCoreAutoConfiguration {
     @Bean
     @ConditionalOnMissingBean(PermissionItsmCallbackService::class)
     fun samplePermissionItsmCallbackService() = SamplePermissionItsmCallbackService()
+
+    @Bean
+    @ConditionalOnMissingBean(PermissionResourceValidateService::class)
+    fun samplePermissionResourceValidateService() = SamplePermissionResourceValidateService()
 }
